requirements.txt
Django==4.0.1
djongo==1.3.6
python-dotenv~=0.15.0
pymongo==4.6.0
pytz~=2020.5

pip~=20.1.1
attrs~=22.1.0
cryptography~=37.0.2
Pillow~=9.3.0
Jinja2~=3.1.2
sqlparse>=0.2.4
asgiref~=3.7.2
colorama~=0.4.4
PyYAML~=6.0
numpy~=1.22.3
click~=7.1.2
certifi~=2020.12.5
requests~=2.25.1
dnspython~=2.4.1
setuptools~=47.1.0
djangorestframework~=3.12.2